requirements.txt
albumentations
ipykernel
jupyter
matplotlib
torch
torchvision
pyyaml
scikit-learn